+++ src/Parser/Evaluated/MobileNaturalParser.php
@@ -7,8 +7,10 @@ namespace Serps\SearchEngine\Google\Parser\Evaluated;
 
 use Serps\SearchEngine\Google\Page\GoogleDom;
 use Serps\SearchEngine\Google\Parser\AbstractParser;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsResultO9g5cc;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsResultZ1m;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsResultZINbbc;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsVideoResult;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ComposedTopStories;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Divider;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ImageGroup;
@@ -16,6 +18,7 @@ use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\ImageGroupCarousel;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\LargeClassicalResult;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\KnowledgeCard;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\Classical\ClassicalCardsResult;
+use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\MapMobile;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\PeopleAlsoAsk;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\SearchResultGroup;
 use Serps\SearchEngine\Google\Parser\Evaluated\Rule\Natural\TweetsCarouselZ1m;
@@ -35,9 +38,12 @@ class MobileNaturalParser extends AbstractParser
         return [
             new Divider(),
             new SearchResultGroup(),
-            new ClassicalCardsResultZINbbc(),
+            new ClassicalCardsResultO9g5cc(),
+            new ClassicalCardsResultZINbbc(), // TODO maybe outdated
             new ClassicalCardsResultZ1m(), // TODO remove (outdated)
+            new ClassicalCardsVideoResult(),
             new ClassicalCardsResult(),
+            new MapMobile(),
             new TweetsCarouselZ1m(), // TODO replace and remove (outdated)
             new ImageGroupCarousel(),
             new ComposedTopStories(),