+++ src/Parser/Evaluated/Rule/Natural/Classical/ClassicalResult.php
@@ -32,25 +32,37 @@ class ClassicalResult implements ParsingRuleInterface
 
         // find the title/url
         /* @var $aTag \DOMElement */
-        $aTag=$dom
-            ->xpathQuery("descendant::h3[@class='r'][1]/a", $node)
+        $aTag = $dom
+            ->xpathQuery("descendant::*[(self::div or self::h3) and @class='r'][1]/a", $node)
             ->item(0);
         if (!$aTag) {
-            throw new InvalidDOMException('Cannot parse a classical resulst.');
+            throw new InvalidDOMException('Cannot parse a classical result.');
+        }
+
+        /* @var $h3Tag \DOMElement */
+        $h3Tag = $dom
+            ->xpathQuery('descendant::h3', $node)
+            ->item(0);
+        if (!$h3Tag) {
+            throw new InvalidDOMException('Cannot parse a classical result.');
         }
 
         $destinationTag = $dom
-            ->cssQuery('div.f>cite', $node)
-            ->item(0);
+            ->cssQuery('div.f cite, div.TbwUpd cite', $node)
+            ->getNodeAt(0);
+
+        if (is_a($destinationTag, Serps\Core\Dom\NullDomNode::class)) {
+            throw new InvalidDOMException('Cannot parse a classical result.');
+        }
 
         $descriptionTag = $dom
             ->xpathQuery("descendant::span[@class='st']", $node)
             ->item(0);
 
         return [
-            'title'   => $aTag->nodeValue,
+            'title'   => $h3Tag->nodeValue,
             'url'     => $dom->getUrl()->resolveAsString($aTag->getAttribute('href')),
-            'destination' => $destinationTag ? $destinationTag->nodeValue : null,
+            'destination' => $destinationTag->getNodeValue(),
             // trim needed for mobile results coming with an initial space
             'description' => $descriptionTag ? trim($descriptionTag->nodeValue) : null,
             'isAmp' => function () use ($dom, $node) {